Add validate role for jsonschema validation of input data
This commits adds the `validate` role, which contains jsonschemas for the
vlan configuration in `group_vars/all/vlans.yaml` and the per-host
interface configuration as well as tasks for performing data validation.

The schemas cover most of but not all of the validations performed by
the erstwhile `validation.py` script. In particular, the schemas do not
forbid mixing L2  configuration (`tagged_vlans`, `untagged_vlan`) with L3
configurastion (`ip4`, `ip6`).

An example of using this role might look like:

    - hosts: localhost
      gather_facts: false
      vars_files:
        - host_vars/OCT7-SW-TORS/interfaces.yml
      roles:
        - role: validate
          vars:
            schema_path: vlan_configuration.yaml
            data: "{{ vlans }}"
        - role: validate
          vars:
            schema_path: interface_configuration.yaml
            data: "{{ interfaces }}"